mood_analyzer/requirements.txt
Flask==2.1.0
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
python-dotenv==0.19.1
